# bloom-wallet-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "bloom-wallet 接口文档 认证方式：在 Authorize 中填写 \"Bearer {token}\"",
        "title": "bloom-wallet",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "返回 token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/project/all": {
            "get": {
                "description": "获取所有项目合约",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salecontract"
                ],
                "summary": "获取所有项目合约",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ProjectContract"
                            }
                        }
                    }
                }
            }
        },
        "/project/{id}": {
            "get": {
                "description": "根据ID获取项目合约",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salecontract"
                ],
                "summary": "根据ID获取项目合约",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目合约ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ProjectContract"
                        }
                    }
                }
            },
            "put": {
                "description": "更新项目合约",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "salecontract"
                ],
                "summary": "更新项目合约",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目合约ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "项目合约更新参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ProjectContractUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/test/ping": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回 pong",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ping"
                ],
                "summary": "ping 测试",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/test/time": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回当前服务器时间（RFC3339 格式）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "获取当前时间",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wallet/account/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查看本地保存的全部账户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "查看所有账户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountsResult"
                        }
                    }
                }
            }
        },
        "/wallet/account/current": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查看当前选中的账户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "查看当前账户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    }
                }
            }
        },
        "/wallet/account/switch": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "按 index 切换当前账户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "切换当前账户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账户 index",
                        "name": "index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    }
                }
            }
        },
        "/wallet/backup": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "备份当前钱包信息（具体内容取决于实现）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "备份钱包",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    }
                }
            }
        },
        "/wallet/create/mnemonic": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建一个助记词",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "创建助记词",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletResult"
                        }
                    }
                }
            }
        },
        "/wallet/create/wallet": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建一个钱包,如果有助记词根据助记词创建,否者创建一个新钱包并返回助记词/地址/密钥等信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "创建钱包",
                "parameters": [
                    {
                        "type": "string",
                        "description": "助记词（空格分隔）；不传则自动生成",
                        "name": "mnemonic",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    }
                }
            }
        },
        "/wallet/import": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传钱包 JSON 文件导入钱包并返回钱包信息",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "导入钱包",
                "parameters": [
                    {
                        "type": "file",
                        "description": "钱包 JSON 文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.WalletResult"
                        }
                    }
                }
            }
        },
        "/wallet/restore": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "从默认备份源恢复钱包（具体内容取决于实现）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "恢复钱包",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    }
                }
            }
        },
        "/wallet/restore/mnemonic": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "通过助记词恢复钱包并返回钱包信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "根据助记词恢复钱包",
                "parameters": [
                    {
                        "type": "string",
                        "description": "助记词（空格分隔）",
                        "name": "mnemonic",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    }
                }
            }
        },
        "/wallet/restore/privateKey": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "通过私钥恢复钱包并返回钱包信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "根据私钥恢复钱包",
                "parameters": [
                    {
                        "type": "string",
                        "description": "私钥（hex 字符串）",
                        "name": "privateKey",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WalletAccountResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ProjectContract": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "金额",
                    "type": "string"
                },
                "amountOfTokensToSell": {
                    "description": "计划出售的代币数量",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "currentPrice": {
                    "description": "当前价格",
                    "type": "integer"
                },
                "decimals": {
                    "description": "代币单位",
                    "type": "integer"
                },
                "description": {
                    "description": "项目说明",
                    "type": "string"
                },
                "earningsWithdrawn": {
                    "description": "是否提取收益",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "description": "项目图标",
                    "type": "string"
                },
                "leftoverWithdrawn": {
                    "description": "是否提取剩余代币",
                    "type": "boolean"
                },
                "maxParticipation": {
                    "description": "单人最大可购代币数量",
                    "type": "string"
                },
                "name": {
                    "description": "项目名称",
                    "type": "string"
                },
                "numberOfRegistrations": {
                    "description": "注册人数",
                    "type": "integer"
                },
                "projectWebsite": {
                    "description": "项目网站",
                    "type": "string"
                },
                "registrationTimeEnds": {
                    "description": "注册结束时间",
                    "type": "string"
                },
                "registrationTimeStarts": {
                    "description": "注册开始时间",
                    "type": "string"
                },
                "saleContractAddress": {
                    "description": "合约地址",
                    "type": "string"
                },
                "saleEnd": {
                    "description": "销售结束时间",
                    "type": "string"
                },
                "saleStart": {
                    "description": "销售开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "symbol": {
                    "description": "代币缩写",
                    "type": "string"
                },
                "tge": {
                    "description": "代币解锁时间",
                    "type": "string"
                },
                "tokenAddress": {
                    "description": "代币地址",
                    "type": "string"
                },
                "tokenName": {
                    "description": "代币名称",
                    "type": "string"
                },
                "tokenPriceInPT": {
                    "description": "代币价格",
                    "type": "string"
                },
                "totalRaised": {
                    "description": "筹集总额",
                    "type": "string"
                },
                "totalTokensSold": {
                    "description": "已卖出的代币数量",
                    "type": "string"
                },
                "unlockTime": {
                    "description": "用户可领取代币的最早时间",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "model.WalletAccount": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "mnemonic": {
                    "description": "仅用于接口返回（wallet.json 的 accounts 节点不会写入）",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                }
            }
        },
        "request.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "request.ProjectContractUpdateRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "paymentToken": {
                    "type": "string"
                },
                "registrationEnd": {
                    "type": "string"
                },
                "registrationStart": {
                    "type": "string"
                },
                "saleAddress": {
                    "type": "string"
                },
                "saleEndTime": {
                    "type": "string"
                },
                "saleOwner": {
                    "type": "string"
                },
                "saleStartTime": {
                    "type": "string"
                },
                "saleToken": {
                    "type": "string"
                },
                "tokenPriceInPT": {
                    "type": "string"
                },
                "totalTokens": {
                    "type": "string"
                },
                "unlockTime": {
                    "type": "string"
                }
            }
        },
        "response.WalletAccountResult": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "mnemonic": {
                    "description": "仅用于接口返回（wallet.json 的 accounts 节点不会写入）",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                }
            }
        },
        "response.WalletAccountsResult": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.WalletAccount"
                    }
                }
            }
        },
        "response.WalletResult": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "mnemonic": {
                    "description": "助记词",
                    "type": "string"
                },
                "privateKey": {
                    "description": "私钥",
                    "type": "string"
                },
                "publicKey": {
                    "description": "公钥",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}